Upgraded numpy

+++ test/test_data.py
@@ -90,7 +90,7 @@ def test_data_type():
     d_float = np.random.randn(3, 50)
     data.set_data(d_float, dim_order='ps')
     assert data.data_type is np.float64, 'Data type did not change.'
-    assert issubclass(type(data.to_numpy()[0, 0, 0]), np.float), (
+    assert issubclass(type(data.to_numpy()[0, 0, 0]), float), (
         'Data type is not a float.')
 
     """
